+++ metropolis-core/tests/legs.rs
@@ -209,9 +209,15 @@ fn legs_test() {
         &preprocess_data.network.get_road_network().unwrap(),
     );
     let weights = NetworkWeights::new(Some(rn_weights));
-    let results =
-        metropolis_core::simulation::run_iteration(weights, None, None, 1, &preprocess_data)
-            .unwrap();
+    let results = metropolis_core::simulation::run_iteration(
+        weights,
+        None,
+        None,
+        1,
+        &preprocess_data,
+        &mut Default::default(),
+    )
+    .unwrap();
     let agent_results = &results.iteration_results.agent_results()[agent_index(0)];
 
     // Departure time from origin: 0.