+++ metropolis-core/tests/spillback.rs
@@ -129,9 +129,15 @@ fn spillback_test() {
         &preprocess_data.network.get_road_network().unwrap(),
     );
     let weights = NetworkWeights::new(Some(rn_weights));
-    let results =
-        metropolis_core::simulation::run_iteration(weights, None, None, 1, &preprocess_data)
-            .unwrap();
+    let results = metropolis_core::simulation::run_iteration(
+        weights,
+        None,
+        None,
+        1,
+        &preprocess_data,
+        &mut Default::default(),
+    )
+    .unwrap();
     let agent_results = results.iteration_results.agent_results();
 
     // Departure times: 0, 7, 5, 6, 9, 30.